version bump to 0.3.0
