+++ uastreamer/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "uastreamer"
-version = "0.2.3"
+version = "0.3.0"
 edition = "2024"
 
 [dependencies]
